madgine

+++ Assets/_Project/Scripts/Gameplay/Buildings/Building/House/HouseGradeData.cs
@@ -8,6 +8,6 @@ namespace _Project.Scripts.Gameplay.Buildings
         public int GradePrice;
         public int MaxUnitsCount;
         public int MaxDurability;
-        public float UnitCost;
+        public int UnitCost;
     }
 }
